@@ -3,40 +3,42 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
+    "ssl": "ng serve --ssl true --ssl-cert '../server.crt' --ssl-key '../server.key'",
     "build": "ng build"
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/schematics": "^9.0.7",
-    "@angular/animations": "^9.0.7",
-    "@angular/cdk": "^9.1.3",
-    "@angular/common": "^9.0.7",
-    "@angular/compiler": "^9.0.7",
-    "@angular/core": "^9.0.7",
-    "@angular/fire": "^5.4.2",
-    "@angular/forms": "^9.0.7",
-    "@angular/material": "^9.1.3",
-    "@angular/platform-browser": "^9.0.7",
-    "@angular/platform-browser-dynamic": "^9.0.7",
-    "@angular/router": "^9.0.7",
-    "@angular/service-worker": "^9.0.7",
+    "@angular/animations": "^10.0.6",
+    "@angular/cdk": "^10.1.1",
+    "@angular/common": "^10.0.6",
+    "@angular/compiler": "^10.0.6",
+    "@angular/core": "^10.0.6",
+    "@angular/fire": "^6.0.2",
+    "@angular/forms": "^10.0.6",
+    "@angular/material": "^10.1.1",
+    "@angular/platform-browser": "^10.0.6",
+    "@angular/platform-browser-dynamic": "^10.0.6",
+    "@angular/pwa": "^0.1000.4",
+    "@angular/router": "^10.0.6",
+    "@angular/service-worker": "^10.0.6",
     "@ngxs/storage-plugin": "^3.6.2",
     "@ngxs/store": "^3.6.2",
-    "hammerjs": "^2.0.8",
-    "rxjs": "^6.5.4",
-    "zone.js": "~0.10.2"
+    "firebase": "^7.17.1",
+    "npm": "^6.14.7",
+    "rxjs": "^6.6.0",
+    "zone.js": "^0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "^0.900.7",
-    "@angular-devkit/build-angular": "~0.900.7",
-    "@angular/cli": "^9.0.7",
-    "@angular/compiler-cli": "^9.0.7",
-    "@angular/language-service": "9.0.7",
-    "@types/node": "^13.9.2",
-    "codelyzer": "^5.2.1",
-    "tslib": "^1.11.1",
-    "tslint": "^6.1.0",
-    "typescript": "^3.7.5"
+    "@angular-devkit/architect": "^0.1000.4",
+    "@angular-devkit/build-angular": "^0.1000.4",
+    "@angular/cli": "^10.0.4",
+    "@angular/compiler-cli": "^10.0.6",
+    "@angular/language-service": "10.0.6",
+    "@types/hammerjs": "^2.0.36",
+    "@types/node": "^14.0.27",
+    "codelyzer": "^6.0.0",
+    "tslib": "^2.0.0",
+    "tslint": "^6.1.2",
+    "typescript": "^3.9.7"
   }
 }
