@@ -1,7 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 
+import { MESSAGE, Auth } from '@dbase/data/data.define';
 import { DataService } from '@dbase/data/data.service';
 import { StateService } from '@dbase/state/state.service';
+
 import { NavigateService } from '@route/router/navigate.service';
 
 @Component({
@@ -11,9 +13,10 @@ import { NavigateService } from '@route/router/navigate.service';
 })
 export class AppComponent implements OnInit {
 	public app$ = this.state.getMemberData();
+	public MESSAGE = MESSAGE;
+	public ROLE = Auth.ROLE;
 
 	constructor(readonly data: DataService, readonly state: StateService, readonly navigate: NavigateService) { }
 
 	ngOnInit() { }
-
 }
