@@ -1,9 +1,9 @@
 import { NgModule, Injector } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
+import { BrowserModule, HammerModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { ServiceWorkerModule } from '@angular/service-worker';
 
-import { RoutingModule } from '@route/route.module';
+import { RoutingModule } from '@route/router/route.module';
 import { DBaseModule } from '@dbase/dbase.module';
 import { AuthModule } from '@service/auth/auth.module';
 import { MaterialModule } from '@service/material/material.module';
@@ -12,25 +12,22 @@ import { AppComponent } from './app.component';
 import { NgxsModule } from '@ngxs/store';
 import { NgxsStoragePluginModule } from '@ngxs/storage-plugin';
 
+import { SafePipe } from '@service/material/safe.pipe';
+
 import { StoreStorage } from '@dbase/sync/sync.define';
 import { environment } from '../environments/environment';
-
-import { InfoSnackBar, WarnSnackBar, ErrorSnackBar } from '@service/material/snack.service';
-import { InfoDialogComponent } from '@service/material/dialog.service';
-import { ForumComponent } from './route/forum/forum.component';
-import { SafePipe } from '@service/material/safe.pipe';
 
 @NgModule({
 	declarations: [
 		AppComponent,
-		ForumComponent,
 		SafePipe,
 	],
 	imports: [
 		NgxsModule.forRoot([], { developmentMode: !environment.production }),
 		NgxsStoragePluginModule.forRoot({ key: StoreStorage }),
-		ServiceWorkerModule.register('ngsw-worker.js', { enabled: environment.production }),
+		// ServiceWorkerModule.register('ngsw-worker.js', { enabled: environment.production }),
 		BrowserModule,
+		HammerModule,
 		DBaseModule,
 		AuthModule,
 		RoutingModule,
@@ -39,7 +36,6 @@ import { SafePipe } from '@service/material/safe.pipe';
 	],
 	providers: [],
 	bootstrap: [AppComponent],
-	entryComponents: [InfoSnackBar, WarnSnackBar, ErrorSnackBar, InfoDialogComponent],
 })
 export class AppModule {
 	static injector: Injector;
