@@ -1,14 +1,12 @@
-import { UserInfo, firestore } from 'firebase/app';
+import { firestore, UserInfo } from 'firebase';
 
-import { COLLECTION, STORE, FIELD, CLASS, EVENT, CONNECT, BONUS, REACT, Auth, PLAN, PRICE, PAYMENT, PROFILE, STATUS, SCHEDULE, MESSAGE, SPAN } from '@dbase/data/data.define';
-import { ISummary } from '@dbase/state/state.define';
-import { getSlice } from '@dbase/state/state.library';
-
-import { TString } from '@lib/type.library';
+import { COLLECTION, STORE, FIELD, CLASS, EVENT, CONNECT, BONUS, REACT, Auth, PLAN, PRICE, PAYMENT, PROFILE, STATUS, SCHEDULE, MESSAGE, SPAN, COLOR, Zoom } from '@dbase/data/data.define';
+import { TString } from '@library/type.library';
+import { Instant } from '@library/instant.library';
 
 type TStoreConfig = STORE.schema | STORE.config | STORE.default;
 type TStoreClient = STORE.class | STORE.event | STORE.price | STORE.plan | STORE.provider | STORE.schedule | STORE.calendar | STORE.location | STORE.instructor | STORE.bonus | STORE.span | STORE.alert | STORE.icon;
-type TStoreUser = STORE.profile | STORE.payment | STORE.gift | STORE.message | STORE.migrate | STORE.attend | STORE.register | STORE.status;
+type TStoreUser = STORE.profile | STORE.payment | STORE.gift | STORE.message | STORE.migrate | STORE.attend | STORE.register | STORE.status | STORE.import;
 type TStoreForum = STORE.comment | STORE.react;
 
 export type FBoolean = boolean | undefined | firestore.FieldValue;
@@ -26,7 +24,7 @@ export interface IMeta {
 	[FIELD.hidden]?: FBoolean;						// valid value, but not to be displayed to the client
 	[FIELD.disable]?: FBoolean;						// valid value, greyed-out to the client
 
-	[FIELD.id]: string;										// the id field on the remote database, generated by FireBase
+	[FIELD.id]: string;										// the id of the document, generated by FireBase
 	[FIELD.store]: STORE;									// each document must declare a 'store' field
 	[FIELD.type]?: string | number;				// an optional 'type' code to qualify the 'store'
 	[FIELD.note]?: TString;								// an optional 'note' on all documents
@@ -34,11 +32,11 @@ export interface IMeta {
 
 /**
  * We have two main types of 'store' documents.  
- * One for 'client' & 'local', keyed by 'store/type?/key',  
- * the other for 'admin', 'member', 'forum' & 'attend', keyed by 'store/type?/uid'
+ * a) one for 'client' & 'local' (keyed by 'store/type?/key'),
+ * b) the other for 'admin', 'member', 'forum' & 'attend' (keyed by 'store/type?/uid')
  */
-export type TStoreBase = IClientBase | IUserBase | IForumBase | IMigrate;
-interface IClientBase extends IMeta {
+export type TStoreBase = IClientBase | IUserBase | IForum | IMigrate;
+export interface IClientBase extends IMeta {
 	[FIELD.store]: TStoreClient | TStoreConfig;
 	[FIELD.key]: string | number;
 	[FIELD.image]?: string;									// an optional icon for the UI
@@ -48,7 +46,7 @@ interface IUserBase extends IMeta {				// this is the base for Member-related do
 	[FIELD.uid]: string;
 	[FIELD.stamp]: number;
 }
-export interface IForumBase extends IMeta {
+export interface IForum extends IMeta {
 	[FIELD.store]: TStoreForum;
 	[FIELD.type]: STORE;										// allow for Forum on any Store type
 	[FIELD.uid]: string;										// the Member making the feedback
@@ -56,26 +54,30 @@ export interface IForumBase extends IMeta {
 	[FIELD.stamp]: number;
 	track: {																// to use in feedback-analysis
 		date: number;													// yearMonthDay
-		[key: string]: string | number;				// additional info to assist tracking
-	},
+	} & Record<string, string | number>,		// additional info to assist tracking
+}
+
+export interface IImport extends IUserBase {
+	[FIELD.store]: STORE.import;						// record of Google Sheet on migrate
+	[FIELD.uid]: string;
+	credit: number;
+	email: string;
+	picture: string;
+	plan: string;
+	sheetName: string;
+	userName: string;
+	providers: TProviderInfo[]
 }
 export interface IMigrate extends IMeta {	// these allow for reconciliation of migrated event
 	[FIELD.store]: STORE.migrate;
 	[FIELD.type]: STORE.event | STORE.class;
 	[FIELD.uid]: string;
 	[FIELD.key]: string;
-	attend: {
+	[COLLECTION.attend]: {
 		[order: string]: CLASS;
 	}
 }
-export interface IStoreMeta extends IMeta {
-	[FIELD.store]: STORE;
-	[key: string]: any;											// add in index-signature
-}
-export type TStoreMeta = IStoreMeta | IStoreMeta[];
-// client documents have a '<key>' field, user documents have a '<uid>' field
-export const isClientDocument = (document: TStoreBase): document is IClientBase =>
-	getSlice(document[FIELD.store]).toString() === COLLECTION.client || getSlice(document[FIELD.store]).toString() === STORE.local;
+export interface IStoreMeta extends IMeta, Record<string, any> { }
 
 //	/client/_default_
 export interface IDefault extends IClientBase {
@@ -94,9 +96,9 @@ export interface ISchema extends IClientBase {
 	[FIELD.store]: STORE.schema;
 	[FIELD.type]: COLLECTION;
 	[FIELD.key]: STORE;
+	[FIELD.sort]?: TString;
 	desc: string;
 	filter?: TString;
-	sort?: TString;
 }
 
 //	/client/price
@@ -115,7 +117,7 @@ export interface IBonus extends IClientBase {
 	level: number;											// number of classes required to attend previously, before this Bonus
 	rule: TString;
 	desc: TString;
-	sort?: number;
+	seq: number;												// sequence of processing
 	count?: number;
 	amount?: number;										// an optional amount to apply
 }
@@ -124,9 +126,9 @@ export interface IBonus extends IClientBase {
 export interface IPlan extends IClientBase {
 	[FIELD.store]: STORE.price;
 	[FIELD.key]: PLAN;
-	bonus?: boolean;										// allowed to apply Bonus
+	[FIELD.sort]: number;
+	[STORE.bonus]?: boolean;						// allowed to apply Bonus
 	desc: string;
-	sort: number;
 	rule?: TString;
 	expiry?: number;										// number of months a Payment is active (before it is auto-expired)
 }
@@ -136,7 +138,7 @@ export interface IClass extends IClientBase {
 	[FIELD.store]: STORE.class;
 	[FIELD.type]: SPAN;
 	[FIELD.key]: CLASS;
-	color: string;
+	color: COLOR;
 	desc?: string;
 }
 
@@ -156,30 +158,30 @@ export interface ICalendar extends IClientBase {
 	[FIELD.key]: number;
 	name: string;
 	desc?: string;
-	day: number;
-	location?: string;
-	instructor?: string;
+	day: Instant.WEEKDAY;
 	start?: string;
+	[STORE.location]?: string;
+	[STORE.instructor]?: string;
 }
 
 //	/client/schedule
 export interface TBonus {							// a sub-type of IBonus
 	[FIELD.id]: string;
 	[FIELD.type]: BONUS;
+	[STORE.gift]?: IGift[];							// the /member/gift updates for this payment
 	count: number;											// the number of this Gift
 	desc: TString;											// message to Member
-	gift?: IGift[];											// the /member/gift updates for this payment
 	amount?: number;										// an optional amount (if other than $0)
 }
 export type TForum = {
-	comment?: TString;
-	react?: REACT;
+	[STORE.comment]?: TString;
+	[STORE.react]?: REACT;
 }
 export interface ISchedule extends IClientBase {
 	[FIELD.store]: STORE.schedule | STORE.calendar;
 	[FIELD.type]: SCHEDULE;
 	[FIELD.key]: CLASS;
-	day: number;
+	day: Instant.WEEKDAY;
 	location?: string;
 	instructor?: string;
 	start: string;
@@ -189,7 +191,7 @@ export interface ISchedule extends IClientBase {
 	bonus?: TBonus;											// the Bonus tracking which can be applied to this Schedule
 	elect?: BONUS;											// name the Bonus the Member chooses (override calc)
 	count?: number;											// number of Attends
-	forum?: TForum;
+	forum?: TForum;											// Member comment to use on their Attend of this Schedule
 }
 
 //	/client/location
@@ -205,9 +207,7 @@ export interface ILocation extends IClientBase {
 		state?: string;
 		postcode?: number;
 	},
-	contact?: {
-		[key: string]: any;
-	},
+	contact?: Record<string, any>,
 	map?: {
 		geo?: {
 			latitude: number;
@@ -222,7 +222,7 @@ export interface IInstructor extends IClientBase {
 	[FIELD.store]: STORE.instructor;
 	name: string;
 	link?: {
-		type: string;
+		[FIELD.type]: string;
 		[property: string]: any;
 	}[];
 }
@@ -239,7 +239,7 @@ export interface ISpan extends IClientBase {
 export interface IAlert extends IClientBase {
 	[FIELD.store]: STORE.alert;
 	[FIELD.type]: STORE.schedule | STORE.event;
-	location?: string;
+	[STORE.location]?: string;
 }
 
 //	/client/react									// Member reaction icons
@@ -247,8 +247,8 @@ export interface IIcon extends IClientBase {
 	[FIELD.store]: STORE.icon;
 	[FIELD.type]: STORE.react | STORE.class | STORE.event | STORE.provider | STORE.bonus;
 	[FIELD.key]: keyof typeof REACT | CLASS | EVENT | Auth.PROVIDER | BONUS;
+	[FIELD.sort]: number;
 	image: string;
-	sort: number;
 }
 
 //	/client/schedule
@@ -323,6 +323,14 @@ export interface IProfilePref extends IProfile {
 	[FIELD.type]: PROFILE.pref;
 	[PROFILE.pref]: TString;
 }
+export interface IProfileToken extends IProfile {
+	[FIELD.type]: PROFILE.token;
+	messaging: {
+		token: string;
+		valid: boolean;
+		error?: { code: string; message: string; };
+	}[]
+}
 export type TProfileInfo = IProfileInfo | IProfileInfo[];
 
 export interface IMemberInfo {				// Conformed Info across Providers
@@ -388,33 +396,33 @@ export interface IAttend extends IUserBase {
 	},
 	track: {													// to use in bonus-checking, attend-analysis, etc.
 		date: number;										// yearMonthDay attended
-		day: number;										// weekDay attended (1-7, 1=Mon)
+		day: Instant.WEEKDAY;						// weekDay attended (1-7, 1=Mon)
 		week: number;										// yearWeek attended
 		month: number;									// yearMonth attended
 	}
 }
 
 //	/forum/comment
-export interface IComment extends IForumBase {
+export interface IComment extends IForum {
 	[FIELD.store]: STORE.comment;
-	comment: TString;
+	[STORE.comment]: TString;
 	response?: {
-		uid: string;
-		stamp: number;
-		comment: TString;
+		[FIELD.uid]: string;
+		[FIELD.stamp]: number;
+		[STORE.comment]: TString;
 		seen?: boolean;
 	}[];
 }
 //	/forum/react
-export interface IReact extends IForumBase {
+export interface IReact extends IForum {
 	[FIELD.store]: STORE.react;
-	react: REACT;
+	[STORE.react]: REACT;
 }
 
 //	/admin/register
 export interface ICustomClaims {		// a special sub-set of fields from the User Token
 	alias?: string;
-	roles?: Auth.ROLE[];
+	roles?: Auth.ROLE[];							// TODO: does this need to be Array?
 	allow?: string[];
 	deny?: string[];
 }
@@ -437,16 +445,17 @@ type TProviderInfo = {
 export interface IRegister extends IUserBase {
 	[FIELD.store]: STORE.register;
 	user: TUser;
-	migrate?: {
-		credit: number;
-		email: string;
-		picture: string;
-		plan: string;
-		sheetName: string;
-		uid: string;
-		userName: string;
-		providers: TProviderInfo[]
-	};
+}
+
+export interface ISummary {
+	bank: number;													// rollover from previous Payment
+	paid: number;													// topUp amount for active Payment
+	adjust: number;												// debit amount for active Payment
+	spend: number;												// sum of Attends' amount
+	funds: number;												// active Payment credit:	bank + paid + adjust - spend
+
+	credit: number;												// whole Account credit:	bank + paid + adjust - spend + pend
+	pend: number;													// sum of not-yet-Active Payments
 }
 
 //	/member/status
@@ -462,4 +471,214 @@ export interface IStatusConnect extends IStatus {
 	[FIELD.type]: STATUS.connect;
 	state: CONNECT;
 	device?: string | null;
+}
+
+export interface IZoom<T extends TStarted | TEnded | TJoined | TLeft | TStatus> {
+	[FIELD.id]: string;
+	[FIELD.create]: string;
+	[FIELD.update]: string;
+	[FIELD.store]: COLLECTION.zoom;
+	[FIELD.type]: Zoom.EVENT;
+	[FIELD.stamp]: number;
+	eventId: string;								// Cloud Functions event Id
+	hook?: number;									// number of times the webhook has been sent
+	head: Record<string, any>;
+	body: T;
+	track: {
+		date: number;
+		day: Instant.WEEKDAY;
+		week: number;
+		month: number;
+	},
+	white?: IWhite;
+}
+
+export type TLeft = {
+	event: Zoom.EVENT.left;
+	payload: {
+		account_id: string;
+		object: {
+			start_time: Date;
+			topic: string;
+			type: number;
+			id: string;
+			duration: number;
+			uuid: string;
+			host_id: string;
+			participant: {
+				id: string;
+				leave_time: Date;
+				user_id: string;
+				user_name: string;
+			},
+			timezone: 'Australia/Sydney';
+		}
+	}
+}
+
+export type TJoined = {
+	event: Zoom.EVENT.joined;
+	payload: {
+		account_id: string;
+		object: {
+			start_time: Date;
+			topic: string;
+			type: number;
+			id: string;
+			duration: number;
+			uuid: string;
+			host_id: string;
+			participant: {
+				id: string;
+				join_time: Date;
+				user_id: string;
+				user_name: string;
+			},
+			timezone: string;
+		}
+	}
+}
+
+export type TStarted = {
+	event: Zoom.EVENT.started;
+	payload: {
+		account_id: string;
+		object: {
+			id: string;
+			uuid: string;
+			host_id: string;
+			topic: string;
+			type: Zoom.TYPE;
+			start_time: Date;
+			timezone: string;
+			duration: number;
+		}
+	};
+}
+
+export type TEnded = {
+	event: Zoom.EVENT.ended;
+	payload: {
+		account_id: string;
+		object: {
+			id: string;
+			uuid: string;
+			host_id: string;
+			topic: string;
+			type: Zoom.TYPE;
+			start_time: Date;
+			end_time: Date;
+			timezone: string;
+			duration: number;
+		}
+	}
+}
+
+export type TStatus = {
+	event: Zoom.EVENT.status;
+	payload: {
+		account_id: string;
+		object: {
+			date_time: Date;
+			email?: string;
+			id: string;
+			presence_status: Zoom.PRESENCE;
+		}
+	}
+}
+
+export interface IMeeting {
+	uuid: string;
+	meeting_id: string;
+	topic?: string;
+	start: {
+		[FIELD.id]: string;
+		[FIELD.stamp]: number;
+		start_time: Date;
+		label: string;							// to be displayed on an UI tab
+		color?: COLOR;
+		white?: IWhite;
+	},
+	end?: {
+		[FIELD.id]: string;
+		[FIELD.stamp]: number;
+		end_time: Date;
+		label: string;							// to be displayed on an UI tab
+	}
+	participants: {
+		participant_id: string;
+		user_id: string;
+		user_name: string;
+		join: {
+			[FIELD.id]: string;
+			[FIELD.stamp]: number;
+			join_time: Date;
+			label: string;							// to be displayed on an UI tab
+			white: IWhite;
+			price?: number;
+			credit?: number;
+			fgcolor?: {
+				alias?: COLOR;
+			}
+			bgcolor?: {
+				price?: COLOR;
+				credit?: COLOR;
+				bonus?: COLOR;
+				priceTip?: string;
+				creditTip?: string;
+				bonusTip?: string;
+			}
+		},
+		leave?: {
+			[FIELD.id]: string;
+			[FIELD.stamp]: number;
+			leave_time: Date;
+			label: string;							// to be displayed on an UI tab
+		}
+	}[]
+}
+
+export interface IZoomEnv {
+	id: string;
+	start: string;
+	end: string;
+	event: string;
+	color: COLOR;
+	day?: number[];
+}
+
+export interface IWhite {
+	class?: string;
+	alias?: string;
+	paid?: true;
+	credit?: number;
+	price?: number;
+	track?: {
+		history: number;
+	},
+	history?: Record<string, any>[];
+	status?: {
+		lastVisit: string;
+		eventType: string;
+		eventPrice: string;
+		eventNote?: string;
+		bonus?: string;
+		hr1?: string;
+		creditExpires: string;
+		creditRemaining: string;
+		trackDaysThisWeek: number;
+		_week: string;
+	},
+	checkIn?: {
+		type: string;
+		price: number;
+		nextAttend: {
+			row: number;
+			col: number;
+			spend: number;
+			bank: number;
+			amt: number;
+			pre: number;
+		}
+	}
 }
