@@ -1,24 +1,21 @@
-import * as firebase from 'firebase/app';
-import { FieldPath } from '@angular/fire/firestore';
-
 import { FIELD } from '@dbase/data/data.define';
+
+export type TWhere = IWhere | IWhere[];
 
 /** Query.where */
 export interface IWhere {
-	fieldPath: string | FieldPath;
-	opStr?: firebase.firestore.WhereFilterOp | '!=';				// this is a *special* to allow additional operator (do not use on Cloud Firestore!)
+	fieldPath: string | firebase.firestore.FieldPath;
+	opStr?: firebase.firestore.WhereFilterOp | '!=';				// this is a *special* to allow additional local operator (do not use on Cloud Firestore queries!)
 	value: any | any[];
 }
-export type TWhere = IWhere | IWhere[];
-
 /** Query.orderBy */
 export interface IOrderBy {
-	fieldPath: string | FieldPath;
+	fieldPath: string | firebase.firestore.FieldPath;
 	directionStr?: firebase.firestore.OrderByDirection;
 }
 /** Collection.Query */
 export interface IQuery {
-	where?: TWhere;
+	where?: IWhere | IWhere[];
 	orderBy?: IOrderBy | IOrderBy[];
 	limit?: number;
 	startAt?: string;
@@ -29,11 +26,11 @@ export interface IQuery {
 
 /** FirestoreFunction getMeta() */
 export interface IDocMeta {
-	exists: boolean;
 	[FIELD.id]: string;
 	[FIELD.create]?: number;
 	[FIELD.update]?: number;
 	[FIELD.access]?: number;
-	subcollections: string[];
+	subcollections?: string[];
+	exists: boolean;
 	path: string;
 }
