@@ -1,14 +1,15 @@
 import { Injectable } from '@angular/core';
 import { State, Action, StateContext, NgxsOnInit } from '@ngxs/store';
-import { SetAdmin, DelAdmin, TruncAdmin, filterState } from '@dbase/state/state.action';
+
+import { Admin, filterState } from '@dbase/state/state.action';
 import { TStateSlice } from '@dbase/state/state.define';
 
 import { FIELD, COLLECTION } from '@dbase/data/data.define';
 import { IStoreMeta } from '@dbase/data/data.schema';
 
-import { asArray } from '@lib/array.library';
-import { cloneObj } from '@lib/object.library';
-import { dbg } from '@lib/logger.library';
+import { asArray } from '@library/array.library';
+import { cloneObj } from '@library/object.library';
+import { dbg } from '@library/logger.library';
 
 /**
  * AdminState is for items that only Admin Users (with roles['admin'] in their customClaims) can see
@@ -23,12 +24,12 @@ export class AdminState implements NgxsOnInit {
 
 	constructor() { this.init(); }
 
-	ngxsOnInit(_ctx: StateContext<TStateSlice<IStoreMeta>>) { this.init(); }
+	ngxsOnInit(_ctx: StateContext<TStateSlice<IStoreMeta>>) { /**this.init(); */}
 
 	private init() { this.dbg('init:'); }
 
-	@Action(SetAdmin)
-	setStore({ getState, setState }: StateContext<TStateSlice<IStoreMeta>>, { payload, debug }: SetAdmin) {
+	@Action(Admin.Set)
+	setStore({ getState, setState }: StateContext<TStateSlice<IStoreMeta>>, { payload, debug }: Admin.Set) {
 		const state = cloneObj(getState()) || {};
 
 		asArray(payload).forEach(doc => {
@@ -44,8 +45,8 @@ export class AdminState implements NgxsOnInit {
 		setState({ ...state });
 	}
 
-	@Action(DelAdmin)
-	delStore({ getState, setState }: StateContext<TStateSlice<IStoreMeta>>, { payload, debug }: DelAdmin) {
+	@Action(Admin.Del)
+	delStore({ getState, setState }: StateContext<TStateSlice<IStoreMeta>>, { payload, debug }: Admin.Del) {
 		const state = cloneObj(getState()) || {};
 
 		asArray(payload).forEach(doc => {
@@ -62,8 +63,8 @@ export class AdminState implements NgxsOnInit {
 		setState({ ...state });
 	}
 
-	@Action(TruncAdmin)
-	truncStore({ setState }: StateContext<TStateSlice<IStoreMeta>>, { debug }: TruncAdmin) {
+	@Action(Admin.Trunc)
+	truncStore({ setState }: StateContext<TStateSlice<IStoreMeta>>, { debug }: Admin.Trunc) {
 		if (debug) this.dbg('truncAdmin');
 		setState({});
 	}
