@@ -1,13 +1,14 @@
 import { Injectable } from '@angular/core';
 import { State, Action, StateContext, NgxsOnInit } from '@ngxs/store';
+
 import { TStateSlice } from '@dbase/state/state.define';
-import { SetAttend, DelAttend, TruncAttend, SyncAttend } from '@dbase/state/state.action';
+import { Attend } from '@dbase/state/state.action';
 
 import { IStoreMeta } from '@dbase/data/data.schema';
 import { FIELD, COLLECTION } from '@dbase/data/data.define';
-import { asArray } from '@lib/array.library';
-import { cloneObj } from '@lib/object.library';
-import { dbg } from '@lib/logger.library';
+import { asArray } from '@library/array.library';
+import { cloneObj } from '@library/object.library';
+import { dbg } from '@library/logger.library';
 
 @Injectable()
 @State<TStateSlice<IStoreMeta>>({
@@ -19,14 +20,14 @@ export class AttendState implements NgxsOnInit {
 
 	constructor() { this.init(); }
 
-	ngxsOnInit(_ctx: StateContext<TStateSlice<IStoreMeta>>) { this.init(); }
+	ngxsOnInit(_ctx: StateContext<TStateSlice<IStoreMeta>>) { /** this.init(); */ }
 
 	private init() {
 		this.dbg('init:');
 	}
 
-	@Action(SetAttend)
-	setStore({ getState, setState, dispatch }: StateContext<TStateSlice<IStoreMeta>>, { payload, debug }: SetAttend) {
+	@Action(Attend.Set)
+	setStore({ getState, setState, dispatch }: StateContext<TStateSlice<IStoreMeta>>, { payload, debug }: Attend.Set) {
 		const state = cloneObj(getState()) || {};
 		let empty: { [segment: string]: boolean; } = {};
 
@@ -43,11 +44,11 @@ export class AttendState implements NgxsOnInit {
 		})
 
 		setState({ ...state });
-		dispatch(new SyncAttend(payload));								// tell any listener we have sync'd
+		dispatch(new Attend.Sync(payload));								// tell any listener we have sync'd
 	}
 
-	@Action(DelAttend)																	// very rare Event
-	delStore({ getState, setState, dispatch }: StateContext<TStateSlice<IStoreMeta>>, { payload, debug }: DelAttend) {
+	@Action(Attend.Del)																	// very rare Event
+	delStore({ getState, setState, dispatch }: StateContext<TStateSlice<IStoreMeta>>, { payload, debug }: Attend.Del) {
 		const state = cloneObj(getState()) || {};
 
 		asArray(payload).forEach(doc => {
@@ -61,11 +62,11 @@ export class AttendState implements NgxsOnInit {
 		})
 
 		setState({ ...state });
-		dispatch(new SyncAttend(payload));									// tell any listener we have sync'd
+		dispatch(new Attend.Sync(payload));									// tell any listener we have sync'd
 	}
 
-	@Action(TruncAttend)
-	truncStore({ setState }: StateContext<TStateSlice<IStoreMeta>>, { debug }: TruncAttend) {
+	@Action(Attend.Trunc)
+	truncStore({ setState }: StateContext<TStateSlice<IStoreMeta>>, { debug }: Attend.Trunc) {
 		if (debug) this.dbg('truncAttend');
 		setState({});
 	}
