@@ -1,5 +1,3 @@
-import * as firebase from 'firebase/app';
-
 import { ICredential } from '@service/auth/auth.interface';
 import { Auth } from '@dbase/data/data.define';
 import { TUser } from '@dbase/data/data.schema';
@@ -13,70 +11,73 @@ export interface IAuthState {
 }
 
 // Actions
-export class CheckSession {
-	static type = '[Auth] CheckSession';
-	constructor(public user: firebase.User | null) { }
+export namespace Login {
+	export class Check {
+		static readonly type = '[Auth] CheckSession';
+		constructor(public user: firebase.User | null) { }
+	}
+	export class Credential {
+		static readonly type = '[Auth] LoginCredential';
+		constructor(public link: ICredential) { }
+	}
+	export class Identity {
+		static readonly type = '[Auth] LoginIdentity';
+		constructor(public authProvider: firebase.auth.AuthProvider, public credential?: any) { }
+	}
+	export class Token {
+		static readonly type = '[Auth] LoginToken';
+		constructor(public token: string, public prefix: Auth.PROVIDER, public user: any) { }
+	}
+	export class OIDC {
+		static readonly type = '[Auth] LoginOIDC';
+		constructor(public token: string, public prefix: Auth.PROVIDER, public user: any) { }
+	}
+	export class Email {
+		static readonly type = '[Auth] LoginEmail';
+		constructor(public email: string, public password: string,
+			public method: 'signIn' | 'createUser' = 'signIn',
+			public credential?: any) { }
+	}
+	export class Anon {
+		static readonly type = '[Auth] LoginAnon';
+	}
+	export class Link {
+		static readonly type = '[Auth] LoginLink';
+		constructor(public link: string, public credential?: any) { }
+	}
+	export class Out {
+		static readonly type = '[Auth] Logout';
+	}
+	export class Off {
+		static readonly type = '[Auth] LogoutSuccess';
+	}
+	export class Other {														// Event to request impersonate another Member
+		static readonly type = '[Auth] AuthOther';
+		constructor(public alias?: string) { }				// 'alias' of Member UID
+	}
 }
-export class LoginCredential {
-	static type = '[Auth] LoginCredential';
-	constructor(public link: ICredential) { }
-}
-export class LoginIdentity {
-	static type = '[Auth] LoginIdentity';
-	constructor(public authProvider: firebase.auth.AuthProvider, public credential?: any) { }
-}
-export class LoginToken {
-	static type = '[Auth] LoginToken';
-	constructor(public token: string, public prefix: Auth.PROVIDER, public user: any) { }
-}
-export class LoginOIDC {
-	static type = '[Auth] LoginOIDC';
-	constructor(public token: string, public prefix: Auth.PROVIDER, public user: any) { }
-}
-export class LoginEmail {
-	static type = '[Auth] LoginEmail';
-	constructor(public email: string, public password: string,
-		public method: 'signIn' | 'createUser' = 'signIn',
-		public credential?: any) { }
-}
-export class LoginAnon {
-	static type = '[Auth] LoginAnon';
-}
-export class LoginLink {
-	static type = '[Auth] LoginLink';
-	constructor(public link: string, public credential?: any) { }
-}
-export class Logout {
-	static type = '[Auth] Logout';
-}
-export class LogoutSuccess {
-	static type = '[Auth] LogoutSuccess';
-}
-
 // Events
-export class LoginRedirect {
-	static type = '[Auth] LoginRedirect';
-}
-export class LoginSuccess {
-	static type = '[Auth] LoginSuccess';
-	constructor(public user: firebase.User) { }
-}
-export class LoginSetup {
-	static type = '[Auth] LoginSetup';
-	constructor(public user: firebase.User) { }
-}
-export class AuthInfo {													// fetch AdditionalUserInfo into auth state
-	static type = '[Auth] AuthInfo';
-	constructor(public info: any) { }
-}
-export class AuthToken {
-	static type = '[Auth] AuthToken';
-}
-export class AuthOther {												// Event to request impersonate another Member
-	static type = '[Auth] AuthOther';
-	constructor(public member: string) { }				// 'member' can be a <uid> or an alias
-}
-export class LoginFailed {
-	static type = '[Auth] LoginFailed';
-	constructor(public error: any) { }
+export namespace Event {
+	export class Redirect {
+		static readonly type = '[Auth] LoginRedirect';
+	}
+	export class Success {
+		static readonly type = '[Auth] LoginSuccess';
+		constructor(public user: firebase.User) { }
+	}
+	export class Setup {
+		static readonly type = '[Auth] LoginSetup';
+		constructor(public user: firebase.User) { }
+	}
+	export class Info {															// fetch AdditionalUserInfo into auth state
+		static readonly type = '[Auth] AuthInfo';
+		constructor(public info: any) { }
+	}
+	export class Token {
+		static readonly type = '[Auth] AuthToken';
+	}
+	export class Failed {
+		static readonly type = '[Auth] LoginFailed';
+		constructor(public error: any) { }
+	}
 }
