@@ -1,16 +1,17 @@
 import { Injectable } from '@angular/core';
 import { State, Action, StateContext, NgxsOnInit, Store } from '@ngxs/store';
-import { SetClient, DelClient, TruncClient, filterState } from '@dbase/state/state.action';
+import { Client, filterState } from '@dbase/state/state.action';
 import { TStateSlice } from '@dbase/state/state.define';
 
-import { SLICES, COMMENT } from '@library/config.define';
-import { setSchema, setConfig } from '@library/config.library';
+import { SLICES, COMMENT } from '@dbase/state/config.define';
+import { setSchema, setConfig } from '@dbase/state/config.library';
 import { STORE, FIELD, COLLECTION } from '@dbase/data/data.define';
 import { IStoreMeta, ISchema, IConfig } from '@dbase/data/data.schema';
 
-import { asArray } from '@lib/array.library';
-import { isEmpty, cloneObj } from '@lib/object.library';
-import { dbg } from '@lib/logger.library';
+import { asArray } from '@library/array.library';
+import { cloneObj } from '@library/object.library';
+import { isEmpty } from '@library/type.library';
+import { dbg } from '@library/logger.library';
 
 @Injectable()
 @State<TStateSlice<IStoreMeta>>({
@@ -22,7 +23,7 @@ export class ClientState implements NgxsOnInit {
 
 	constructor(private readonly store: Store) { this.init(); }
 
-	ngxsOnInit(_ctx: StateContext<TStateSlice<IStoreMeta>>) { this.init(); }
+	ngxsOnInit(_ctx: StateContext<TStateSlice<IStoreMeta>>) { /** this.init(); */ }
 
 	private init() {
 		this.dbg('init:');
@@ -34,8 +35,8 @@ export class ClientState implements NgxsOnInit {
 			})
 	}
 
-	@Action(SetClient)
-	setStore({ getState, setState }: StateContext<TStateSlice<IStoreMeta>>, { payload, debug }: SetClient) {
+	@Action(Client.Set)
+	setStore({ getState, setState }: StateContext<TStateSlice<IStoreMeta>>, { payload, debug }: Client.Set) {
 		const state = cloneObj(getState()) || {};
 		const schema: IStoreMeta[] = [];
 		const config: IStoreMeta[] = [];
@@ -60,8 +61,8 @@ export class ClientState implements NgxsOnInit {
 		setState({ ...state });
 	}
 
-	@Action(DelClient)
-	delStore({ getState, setState }: StateContext<TStateSlice<IStoreMeta>>, { payload, debug }: DelClient) {
+	@Action(Client.Del)
+	delStore({ getState, setState }: StateContext<TStateSlice<IStoreMeta>>, { payload, debug }: Client.Del) {
 		const state = cloneObj(getState()) || {};
 		const schema: IStoreMeta[] = [];
 		const config: IStoreMeta[] = [];
@@ -87,8 +88,8 @@ export class ClientState implements NgxsOnInit {
 		setState({ ...state });
 	}
 
-	@Action(TruncClient)
-	truncStore({ setState }: StateContext<TStateSlice<IStoreMeta>>, { debug }: TruncClient) {
+	@Action(Client.Trunc)
+	truncStore({ setState }: StateContext<TStateSlice<IStoreMeta>>, { debug }: Client.Trunc) {
 		if (debug) this.dbg('truncClient');
 		setState({});
 	}
