@@ -1,13 +1,14 @@
 import { Injectable } from '@angular/core';
 import { State, Action, StateContext, NgxsOnInit } from '@ngxs/store';
+
 import { TStateSlice } from '@dbase/state/state.define';
 import { IStoreMeta } from '@dbase/data/data.schema';
-import { SetMember, DelMember, TruncMember, filterState } from '@dbase/state/state.action';
+import { Member, filterState } from '@dbase/state/state.action';
 
 import { FIELD, COLLECTION } from '@dbase/data/data.define';
-import { asArray } from '@lib/array.library';
-import { cloneObj } from '@lib/object.library';
-import { dbg } from '@lib/logger.library';
+import { asArray } from '@library/array.library';
+import { cloneObj } from '@library/object.library';
+import { dbg } from '@library/logger.library';
 
 @Injectable()
 @State<TStateSlice<IStoreMeta>>({
@@ -19,17 +20,18 @@ export class MemberState implements NgxsOnInit {
 
 	constructor() { this.init(); }
 
-	ngxsOnInit(_ctx: StateContext<TStateSlice<IStoreMeta>>) { this.init(); }
+	ngxsOnInit(_ctx: StateContext<TStateSlice<IStoreMeta>>) { /** this.init(); */ }
 
 	private init() {
 		this.dbg('init:');
 	}
 
-	@Action(SetMember)
-	setStore({ getState, setState }: StateContext<TStateSlice<IStoreMeta>>, { payload, debug }: SetMember) {
+	@Action(Member.Set)
+	setStore({ getState, setState }: StateContext<TStateSlice<IStoreMeta>>, { payload, debug }: Member.Set) {
 		const state = cloneObj(getState()) || {};
 
 		asArray(payload).forEach(doc => {
+			doc
 			const store = doc[FIELD.store];
 
 			state[store] = filterState(state, doc);
@@ -41,8 +43,8 @@ export class MemberState implements NgxsOnInit {
 		setState({ ...state });
 	}
 
-	@Action(DelMember)
-	delStore({ getState, setState }: StateContext<TStateSlice<IStoreMeta>>, { payload, debug }: DelMember) {
+	@Action(Member.Del)
+	delStore({ getState, setState }: StateContext<TStateSlice<IStoreMeta>>, { payload, debug }: Member.Del) {
 		const state = cloneObj(getState()) || {};
 		const segment = FIELD.store;
 
@@ -58,8 +60,8 @@ export class MemberState implements NgxsOnInit {
 		setState({ ...state });
 	}
 
-	@Action(TruncMember)
-	truncStore({ setState }: StateContext<TStateSlice<IStoreMeta>>, { debug }: TruncMember) {
+	@Action(Member.Trunc)
+	truncStore({ setState }: StateContext<TStateSlice<IStoreMeta>>, { debug }: Member.Trunc) {
 		if (debug) this.dbg('truncMember');
 		setState({});
 	}
