@@ -1,11 +1,11 @@
 import { Observable } from 'rxjs';
 
 import { IAuthState } from './auth.action';
-import { STORE } from '@dbase/data/data.define';
+import { STORE, BONUS, COLLECTION } from '@dbase/data/data.define';
 import {
 	IDefault, IProfilePlan, IProfilePref, IPrice, IPlan, IPayment, IAttend, ISchedule, IClass, IEvent, ICalendar,
 	ILocation, IInstructor, IProfileInfo, IStoreMeta, ISpan, IAlert, IMessage, IRegister, ISchema, IConfig, IGift, IBonus,
-	IStatusConnect, IStatusAccount, TBonus, IIcon, IProvider, IReact, IComment,
+	IStatusConnect, IStatusAccount, TBonus, IIcon, IProvider, IReact, IComment, IImport, ISummary,
 } from '@dbase/data/data.schema';
 
 export enum SLICE {
@@ -20,19 +20,19 @@ export enum SLICE {
 }
 
 /**
- * State contains a copy of remote documents, sliced into distinct 'store' objects.  
+ * State contains a copy of remote documents, sliced into distinct 'segment' objects.  
  * For example:  
  * >client: {                                	// Client slice, holds public data  
- * >		class: [ {class documents} ],       	// Class store, holds documents that describe classes  
+ * >	class: [ {class documents} ],       		// Class store, holds documents that describe classes  
  * > 	price: [ {price documents} ],      			// Price store, holds documents that describe pricing structure  
  * >		...  
  * >	},  
  * >member: {																	// Member slice, keyed by User.uid  
- * >     profile: [ {profile documents} ],   	// Plan, Claims, User, etc.  
- * >     payment: [ {payment documents} ],   	// describe payments  
+ * > 	profile: [ {profile documents} ],   		// Plan, Claims, User, etc.  
+ * >	payment: [ {payment documents} ],   		// describe payments  
  * > },  
  * >attend: {																	// Attend slice, keyed by User.uid, linked to /member/payment documents  
- * >     {paymentId}: [ {attendance documents} ],// hold a number of Attendances against a particular account-payment  
+ * >	{paymentId}: [ {attendance documents} ],// hold a number of Attendances against a particular account-payment  
  * }  
  */
 
@@ -46,15 +46,16 @@ export interface IUserState {
 
 export interface IAdminState {
 	[STORE.register]: IRegister[];
-	account: IStatusAccount[];
-	connect: IStatusConnect[];
-	dash: {
+	[STORE.account]: IStatusAccount[];
+	[STORE.connect]: IStatusConnect[];
+	[STORE.import]: IImport[];
+	dash: {																// align elements of each Admin Array by UID
 		[STORE.register]: IRegister;
-		account: IStatusAccount;
-		connect: IStatusConnect;
+		[STORE.account]?: IStatusAccount;
+		[STORE.connect]?: IStatusConnect;
+		[STORE.import]?: IImport;
 	}[]
-};
-
+}
 
 export interface IApplicationState {		// application-wide settings
 	application: {
@@ -65,55 +66,46 @@ export interface IApplicationState {		// application-wide settings
 }
 
 export interface IMemberState extends IUserState, IApplicationState {
-	member: {
-		plan: IProfilePlan[];             	// member's effective plan
+	[COLLECTION.member]: {
+		[STORE.plan]: IProfilePlan[];       // member's effective plan
 		info: IProfileInfo[];             	// array of AdditionalUserInfo documents
 		pref: IProfilePref[];								// member's preferences
-		message: IMessage[];								// array of messages to a Member
-		gift: IGift[];											// array of gifts for a Member
+		[STORE.message]: IMessage[];				// array of messages to a Member
+		[STORE.gift]: IGift[];							// array of gifts for a Member
 	}
 }
 
 export interface IProviderState {
-	client: {
-		provider: IProvider[];							// array of effective Provider documents
-		icon: IIcon[];											// array of Provider icons
+	[COLLECTION.client]: {
+		[STORE.provider]: IProvider[];			// array of effective Provider documents
+		[STORE.icon]: IIcon[];							// array of Provider icons
 	}
 }
 export interface IPlanState extends IMemberState {
-	client: {
-		plan: IPlan[];                     	// array of effective Plan documents
-		price: IPrice[];                   	// array of effective Price documents
-		icon: IIcon[];											// array of Plan icons
+	[COLLECTION.client]: {
+		[STORE.plan]: IPlan[];              // array of effective Plan documents
+		[STORE.price]: IPrice[];						// array of effective Price documents
+		[STORE.icon]: IIcon[];							// array of Plan icons
 	}
 }
 
 export interface IBonusState extends IUserState, IApplicationState {
-	client: {
-		bonus: IBonus[];										// array of Bonus scheme rules
+	[COLLECTION.client]: {
+		[STORE.bonus]: IBonus[];						// array of Bonus scheme rules
 	}
-	member: {
-		gift: IGift[];											// array of gifts to a Member
+	[COLLECTION.member]: {
+		[STORE.gift]: IGift[];							// array of gifts to a Member
 	}
 	track: {
-		gift: IAttend[];										// array of Attends against their gifts
-		week: IAttend[];										// array of Attends against the current week
-		month: IAttend[];										// array of Attends against the current month
+		[BONUS.gift]: IAttend[];						// array of Attends against their gifts
+		[BONUS.week]: IAttend[];						// array of Attends against the current week
+		[BONUS.class]: IAttend[];						// array of Attends against the current week
+		[BONUS.month]: IAttend[];						// array of Attends against the current month
 	}
 }
 
-export interface ISummary {
-	bank: number;													// rollover from previous Payment
-	paid: number;													// topUp Amount
-	pend: number;													// sum of not-yet-Active Payments
-	adjust: number;												// debits
-	spend: number;												// sum of Attends' amount
-
-	credit: number;												// bank + pend + paid + adjust - spend
-	funds: number;												// bank + paid + adjust - spend
-}
 export interface IAccountState extends IMemberState, IPlanState {
-	account: {
+	[STORE.account]: {
 		payment: IPayment[];								// array of open payment documents, sorted by <stamp>
 		attend: IAttend[];
 		summary: ISummary;
@@ -129,33 +121,35 @@ export interface IPaymentState {
 }
 
 export interface ITimetableState extends IMemberState, IApplicationState {
-	client: {
-		schedule?: ISchedule[];
-		class?: IClass[];
-		event?: IEvent[];
-		calendar?: ICalendar[];
-		location?: ILocation[];
-		instructor?: IInstructor[];
-		plan?: IPlan[];
-		price?: IPrice[];
-		span?: ISpan[];
-		alert?: IAlert[];
-		bonus?: IBonus[];
-		icon?: IIcon[];
+	[COLLECTION.client]: {
+		[STORE.schedule]?: ISchedule[];
+		[STORE.class]?: IClass[];
+		[STORE.event]?: IEvent[];
+		[STORE.calendar]?: ICalendar[];
+		[STORE.diary]?: ICalendar[];
+		[STORE.location]?: ILocation[];
+		[STORE.instructor]?: IInstructor[];
+		[STORE.plan]?: IPlan[];
+		[STORE.price]?: IPrice[];
+		[STORE.span]?: ISpan[];
+		[STORE.alert]?: IAlert[];
+		[STORE.bonus]?: IBonus[];
+		[STORE.icon]?: IIcon[];
 	},
-	attend: {
+	[COLLECTION.attend]: {
 		attendGift: IAttend[];
 		attendWeek: IAttend[];
+		attendClass: IAttend[];
 		attendMonth: IAttend[];
 		attendToday: IAttend[];
 	},
-	bonus?: TBonus,
-	forum?: IForumState["forum"];
+	[STORE.bonus]?: TBonus,
+	[COLLECTION.forum]?: IForumState["forum"];
 }
 
 export interface IForumState {
-	forum: {
-		comment?: IComment[];
-		react?: IReact[];
+	[COLLECTION.forum]: {
+		[STORE.comment]?: IComment[];
+		[STORE.react]?: IReact[];
 	}
 }
