@@ -1,7 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 
-import { ForumService } from '../../service/forum/forum.service';
 import { DataService } from '@dbase/data/data.service';
+import { ForumService } from '@service/forum/forum.service';
+import { ReactionService } from '@service/forum/reaction.service';
 
 @Component({
 	selector: 'wb-forum',
@@ -9,7 +10,7 @@ import { DataService } from '@dbase/data/data.service';
 })
 export class ForumComponent implements OnInit {
 
-	constructor(private forum: ForumService, private data: DataService) { }
+	constructor(private forum: ForumService, private react: ReactionService, private data: DataService) { }
 
 	ngOnInit() { }
 
