@@ -10,8 +10,9 @@ import { StateService } from '@dbase/state/state.service';
 import { FIELD, PRICE } from '@dbase/data/data.define';
 import { IPrice, IPlan } from '@dbase/data/data.schema';
 
-import { isUndefined } from '@lib/type.library';
-import { dbg } from '@lib/logger.library';
+import { isUndefined } from '@library/type.library';
+import { padString } from '@library/string.library';
+import { dbg } from '@library/logger.library';
 
 @Component({
 	selector: 'wb-plan',
@@ -21,7 +22,7 @@ export class PlanComponent implements OnInit {
 	public data$!: Observable<IPlanState>;
 	private dbg = dbg(this);
 
-	constructor(private readonly member: MemberService, private readonly state: StateService, public dialog: DialogService) { }
+	constructor(public readonly member: MemberService, private readonly state: StateService, public dialog: DialogService) { }
 
 	ngOnInit() {
 		this.data$ = this.state.getPlanData().pipe(
@@ -56,19 +57,15 @@ export class PlanComponent implements OnInit {
 			.filter(row => row[FIELD.key] === plan[FIELD.key])
 			.forEach(price => {
 				if (price[FIELD.type] === PRICE.topUp && price.amount)
-					content.push(`$${this.fixString(price.amount)} to topUp your account`);
+					content.push(`$${padString(price.amount)} to topUp your account`);
 				if (price[FIELD.type] === PRICE.hold)
-					content.push(`$${this.fixString(price.amount)} to put funds on-hold for 90-days`)
+					content.push(`$${padString(price.amount)} to put funds on-hold for 90-days`)
 				if (price[FIELD.type] === PRICE.full)
-					content.push(`$${this.fixString(price.amount)} for a one-hour Class`);
+					content.push(`$${padString(price.amount)} for a one-hour Class`);
 				if (price[FIELD.type] === PRICE.half)
-					content.push(`$${this.fixString(price.amount)} for a half-hour Class`)
+					content.push(`$${padString(price.amount)} for a half-hour Class`)
 			})
 
 		this.dialog.open({ image, title, subtitle, actions, content });
-	}
-
-	private fixString(nbr: number) {
-		return nbr.toFixed(2).padStart(6, '\u007F');
 	}
 }
