@@ -1,59 +1,60 @@
-import * as firebase from 'firebase/app';
+import { auth } from 'firebase/app';
 
 import { IUserState } from '@dbase/state/state.define';
 import { Auth } from '@dbase/data/data.define';
 
-import { isNull } from '@lib/type.library';
+import { isNull } from '@library/type.library';
 
 interface IEmailToken {
 	email: string;
 	password: string;
 }
 export const getAuthProvider = (providerId: string,
-	token?: (firebase.auth.IdTokenResult & IEmailToken) | null): [Auth.METHOD | undefined, firebase.auth.AuthProvider | undefined, firebase.auth.AuthCredential | undefined] => {
-	let authProvider: firebase.auth.AuthProvider | undefined;
-	let authCredential: firebase.auth.AuthCredential | undefined;
+	token?: (auth.IdTokenResult & IEmailToken) | null): [Auth.METHOD | undefined, auth.AuthProvider | undefined, auth.AuthCredential | undefined] => {
+	let authProvider: auth.AuthProvider | undefined;
+	let authCredential: auth.AuthCredential | undefined;
 	let method: Auth.METHOD | undefined = Auth.METHOD.identity;									// default to 'identity' 
 
 	switch (providerId) {
 		case Auth.METHOD.email:
 			method = Auth.METHOD.email
-			authProvider = new firebase.auth.EmailAuthProvider();
+			authProvider = new auth.EmailAuthProvider();
+			// authProvider = new auth.EmailAuthProvider();
 			if (token)
-				authCredential = firebase.auth.EmailAuthProvider.credential(token.email, token.password);
+				authCredential = auth.EmailAuthProvider.credential(token.email, token.password);
 			break;
 
 		case 'google':
 		case 'google.com':
-			authProvider = new firebase.auth.GoogleAuthProvider();
+			authProvider = new auth.GoogleAuthProvider();
 			if (token)
-				authCredential = firebase.auth.GoogleAuthProvider.credential(token.token);
+				authCredential = auth.GoogleAuthProvider.credential(token.token);
 			break;
 
 		case 'twitter':
 		case 'twitter.com':
-			authProvider = new firebase.auth.TwitterAuthProvider();
+			authProvider = new auth.TwitterAuthProvider();
 			if (token)
-				authCredential = firebase.auth.TwitterAuthProvider.credential(token.token, '');
+				authCredential = auth.TwitterAuthProvider.credential(token.token, '');
 			break;
 
 		case 'github':
 		case 'github.com':
-			authProvider = new firebase.auth.GithubAuthProvider();
+			authProvider = new auth.GithubAuthProvider();
 			if (token)
-				authCredential = firebase.auth.GithubAuthProvider.credential(token.token);
+				authCredential = auth.GithubAuthProvider.credential(token.token);
 			break;
 
 		case 'facebook':
 		case 'facebook.com':
-			authProvider = new firebase.auth.FacebookAuthProvider();
+			authProvider = new auth.FacebookAuthProvider();
 			if (token)
-				authCredential = firebase.auth.FacebookAuthProvider.credential(token.token);
+				authCredential = auth.FacebookAuthProvider.credential(token.token);
 			break;
 
 		// case 'linkedin':
 		// case 'linkedin.com':
-		// 	authProvider = new firebase.auth.OAuthProvider('oidc.' + providerId);
+		// 	authProvider = new auth.OAuthProvider('oidc.' + providerId);
 		// if (token)
 		// authCredential = authProvider.credential(token, null);
 		// 	break;
