@@ -3,7 +3,7 @@ import { map, switchMap, take } from 'rxjs/operators';
 
 import { Store } from '@ngxs/store';
 import { StateService } from '@dbase/state/state.service';
-import { LoginIdentity, Logout, LoginEmail, LoginToken, AuthInfo, LoginAnon } from '@dbase/state/auth.action';
+import { Login, Event } from '@dbase/state/auth.action';
 
 import { AuthModule } from '@service/auth/auth.module';
 import { getAuthProvider, isActive } from '@service/auth/auth.library';
@@ -13,9 +13,9 @@ import { FireService } from '@dbase/fire/fire.service';
 import { FIELD, STORE, Auth } from '@dbase/data/data.define';
 import { IProvider, IConfig } from '@dbase/data/data.schema';
 
-import { asArray } from '@lib/array.library';
-import { getConfig } from '@library/config.library';
-import { dbg } from '@lib/logger.library';
+import { asArray } from '@library/array.library';
+import { getConfig } from '@dbase/state/config.library';
+import { dbg } from '@library/logger.library';
 
 @Injectable({ providedIn: AuthModule })
 export class AuthService {
@@ -45,11 +45,12 @@ export class AuthService {
 	}
 
 	public signOut() {
-		this.store.dispatch(new Logout());
+		this.store.dispatch(new Login.Out());
 	}
 
 	public async signIn(provider: IProvider, opts: Record<string, any> = {}) {
-		this.dbg('signIn: %j', provider);
+		const { image, ...rest } = provider;								// drop the <image> for debug
+		this.dbg('signIn: %j', rest);
 
 		switch (provider[FIELD.type]) {
 			case undefined:
@@ -105,20 +106,20 @@ export class AuthService {
 		if (provider.params)
 			(authProvider as TParams).setCustomParameters(provider.params);
 
-		return this.store.dispatch(new LoginIdentity(authProvider))
+		return this.store.dispatch(new Login.Identity(authProvider))
 	}
 
 	/** This runs in the main thread */
 	private async signInOAuth(provider: IProvider) {
 		const urlQuery = `prefix=${provider.prefix}`;
 		const oauth = await this.state.asPromise(this.state.getCurrent<IConfig>(STORE.config))
-			.then(config => getConfig(config, 'oauth'))
+			.then(config => getConfig(config, STORE.provider, 'oauth'))
 			.then(oauth => oauth.value)
 
 		const child = this.openChannel('token');// link to the child popup
 		child.onmessage = (msg) => {						// when child sends a message...
 			child.close();												// 	close BroadcastChannel
-			this.store.dispatch(new AuthInfo({ info: JSON.parse(msg.data) }));
+			this.store.dispatch(new Event.Info({ info: JSON.parse(msg.data) }));
 		}
 
 		window.open(`${oauth.request_url}?${urlQuery}`, '_blank', 'height=600,width=400');
@@ -133,7 +134,7 @@ export class AuthService {
 		const parent = this.openChannel('token');// link to the parent of this popup
 		this.dbg('signInToken: %j', response);
 
-		return this.store.dispatch(new LoginToken(response.token, response.prefix, response.user))
+		return this.store.dispatch(new Login.Token(response.token, response.prefix, response.user))
 			.pipe(switchMap(_ => this.auth$))			// this will fire multiple times, as AuthState settles
 			.subscribe(state => {
 				if (state.auth.info) 								// tell the parent to update State
@@ -145,15 +146,15 @@ export class AuthService {
 	}
 
 	private signInEmail(provider: IProvider, email: string, password: string) {
-		return this.store.dispatch(new LoginEmail(email, password))
+		return this.store.dispatch(new Login.Email(email, password))
 	}
 
 	private signInAnon(provider: IProvider) {
-		return this.store.dispatch(new LoginAnon())
+		return this.store.dispatch(new Login.Anon())
 	}
 
 	private signInCustom(token: string) {
-		return this.store.dispatch(new LoginToken(token, Auth.PROVIDER.jwt, {}));
+		return this.store.dispatch(new Login.Token(token, Auth.PROVIDER.jwt, {}));
 	}
 
 	private signInOIDC(provider: IProvider) { }
