@@ -4,7 +4,7 @@ import { DBaseModule } from '@dbase/dbase.module';
 import { DataService } from '@dbase/data/data.service';
 import { STORE, REACT } from '@dbase/data/data.define';
 
-import { dbg } from '@lib/logger.library';
+import { dbg } from '@library/logger.library';
 
 @Injectable({ providedIn: DBaseModule })
 export class ReactionService {
