@@ -1,10 +1,11 @@
 import { ITimetableState } from '@dbase/state/state.define';
-import { FIELD, BONUS } from '@dbase/data/data.define';
-import { IGift, TBonus, IBonus } from '@dbase/data/data.schema';
+import { FIELD, BONUS, COLLECTION, STORE } from '@dbase/data/data.define';
+import { IGift, TBonus, IAttend, ISchedule } from '@dbase/data/data.schema';
 
-import { TDate, getDate, Instant } from '@lib/instant.library';
-import { TString, isUndefined } from '@lib/type.library';
-import { asArray } from '@lib/array.library';
+import { TDate, getDate, Instant } from '@library/instant.library';
+import { TString, isUndefined } from '@library/type.library';
+import { asArray } from '@library/array.library';
+import { plural } from '@library/string.library';
 
 /**
  * Determine if provided event is entitled to a Gift or a Bonus  
@@ -15,18 +16,17 @@ import { asArray } from '@lib/array.library';
  */
 export const calcBonus = (source: ITimetableState, event: string, date?: TDate, elect?: BONUS) => {
 	const now = getDate(date);
-	const upd: IGift[] = [];																// an array of updates for caller to apply on Gifts
+	const today = now.format(Instant.FORMAT.yearMonthDay);
+	const upd: IGift[] = [];																// an array of updates for caller to apply on /member/gift
 	let bonus = {} as TBonus;																// calculated Bonus entitlement
 
-	const gifts = source.member.gift;												// the active Gifts for this Member
-	const plan = (source.client.plan || [])[0];							// Member's current plan
-	const { attendGift = [], attendWeek = [], attendMonth = [], attendToday = [] } = source.attend;
-	const scheme = (source.client.bonus || [])
-		.reduce((acc, row) => {
-			acc[row[FIELD.key]] = row;													// get the <rules> for each Bonus
-			return acc;
-		}, {} as Record<BONUS, IBonus>);
+	const gifts = source[COLLECTION.member][STORE.gift];								// the active Gifts for this Member
+	const plan = (source[COLLECTION.client][STORE.plan] || [])[0];			// Member's current plan
+	const { attendGift = [], attendWeek = [], attendMonth = [], attendToday = [] } = source[COLLECTION.attend];
+	const scheme = (source[COLLECTION.client][STORE.bonus] || [])
+		.groupBy<BONUS>(FIELD.key);														// get the <rules> for each Bonus
 
+	// TODO: I am of the opinion that <gift> should be tested *after* other bonus-types.  
 	switch (true) {
 		/**
 		 * Admin adds 'Gift' records to a Member's account which will detail the start-date
@@ -62,7 +62,7 @@ export const calcBonus = (source: ITimetableState, event: string, date?: TDate, 
 						[FIELD.id]: gift[FIELD.id],
 						[FIELD.type]: BONUS.gift,
 						count: gift.count,
-						desc: `${giftLeft} gift${giftLeft > 1 ? 's' : ''} left`,
+						desc: `${giftLeft} ${plural(giftLeft, 'gift')}`,
 					}
 				}
 			});
@@ -74,27 +74,51 @@ export const calcBonus = (source: ITimetableState, event: string, date?: TDate, 
 			break;																							//	so no further checking needed
 
 		/**
-		 * The Week scheme qualifies as a Bonus if the Member attends the required number of non-bonus classes in a week (scheme.week.level).  
+		 * The Week scheme qualifies as a Bonus if the Member attends the required number of non-bonus days in a week (scheme.week.level).  
 		 * (note: even 'Gift Attends' count towards a Bonus)
 		 * The Member must also have claimed less than the free limit (scheme.week.free) to enable this bonus
 		 */
 		case scheme.week
 			&& attendWeek																				// sum the non-Bonus -or- Gift Attends
 				.filter(row => isUndefined(row.bonus) || row.bonus[FIELD.type] === BONUS.gift)
+				.filter(row => row.track[FIELD.date] !== today)		// dont count today's attend
 				.distinct(row => row.track[FIELD.date])						// de-dup by day-of-week
 				.length >= scheme.week.level											// must attend the 'level' number
 			&& attendWeek																				// sum the 'week' bonus claimed
-				.filter(row => row.bonus && row.bonus[FIELD.type] === BONUS.week)
+				.filter(row => row.bonus?.[FIELD.type] === BONUS.week)
 				.length < scheme.week.free												// must attend less than 'free' number
 			&& (isUndefined(elect) || elect === BONUS.week):		// if not skipped, then Bonus will apply
 
 			bonus = {
 				[FIELD.id]: scheme.week[FIELD.id],
 				[FIELD.type]: BONUS.week,
-				count: attendWeek.filter(row => row.bonus && row.bonus[FIELD.type] === BONUS.week).length + 1,
+				count: attendWeek.filter(row => row.bonus?.[FIELD.type] === BONUS.week).length + 1,
 				desc: scheme.week.desc,
 			}
 			break;
+
+		/**
+		 * The Class scheme qualifies as a Bonus if the Member attends the required number of non-bonus classes in a week (scheme.class.level).  
+		 * (note: 'Gift' attends do not count towards a Bonus)
+		 * THe Member must also have claimed less than the free limit (scheme.class.free) to enable this bonus
+		 */
+		case scheme.class
+			&& attendWeek
+				.filter(row => isUndefined(row.bonus) || row.bonus[FIELD.type] === BONUS.gift)	// only non-Bonus attends
+				.length >= scheme.class.level											// must attend the 'level' number
+			&& attendWeek																				// sum the 'week' bonus claimed
+				.filter(row => row.bonus?.[FIELD.type] === BONUS.class)
+				.length < scheme.class.free												// must attend less than 'free' number
+			&& (isUndefined(elect) || elect === BONUS.class):		// if not skipped, then Bonus will apply
+
+			bonus = {
+				[FIELD.id]: scheme.class[FIELD.id],
+				[FIELD.type]: BONUS.class,
+				count: attendWeek.filter(row => row.bonus?.[FIELD.type] === BONUS.class).length + 1,
+				desc: scheme.class.desc,
+			}
+			break;
+
 
 		/**
 		 * The Sunday scheme qualifies as a Bonus if the Member attends the required number of non-bonus classes in a week (scheme.sunday.level).  
@@ -104,6 +128,7 @@ export const calcBonus = (source: ITimetableState, event: string, date?: TDate, 
 		case scheme.sunday
 			&& attendWeek																				// count Attends this week either Gift or non-Bonus
 				.filter(row => isUndefined(row.bonus) || row.bonus[FIELD.type] === BONUS.gift)
+				.filter(row => row.track[FIELD.date] !== today)		// dont count today's attend
 				.distinct(row => row.track[FIELD.date])						// de-dup by day-of-week
 				.length >= scheme.sunday.level										// required number of Attends this week
 			&& now.dow === Instant.WEEKDAY.Sun									// today is 'Sunday'
@@ -113,7 +138,7 @@ export const calcBonus = (source: ITimetableState, event: string, date?: TDate, 
 			bonus = {
 				[FIELD.id]: scheme.sunday[FIELD.id],
 				[FIELD.type]: BONUS.sunday,
-				count: attendWeek.filter(row => row.bonus && row.bonus[FIELD.type] === BONUS.sunday).length + 1,
+				count: attendWeek.filter(row => row.bonus?.[FIELD.type] === BONUS.sunday).length + 1,
 				desc: scheme.sunday.desc,
 			}
 			break;
@@ -125,24 +150,44 @@ export const calcBonus = (source: ITimetableState, event: string, date?: TDate, 
 		case scheme.month
 			&& attendMonth
 				.filter(row => isUndefined(row.bonus) || row.bonus[FIELD.type] === BONUS.gift)
+				.filter(row => row.track[FIELD.date] !== today)		// dont count today's attend
 				.distinct(row => row.track[FIELD.date])
 				.length >= scheme.month.level
 			&& attendMonth
-				.filter(row => row.bonus && row.bonus[FIELD.type] === BONUS.month)
+				.filter(row => row.bonus?.[FIELD.type] === BONUS.month)
 				.length < scheme.month.free
 			&& (isUndefined(elect) || elect === BONUS.month):
 
 			bonus = {
 				[FIELD.id]: scheme.month[FIELD.id],
 				[FIELD.type]: BONUS.month,
-				count: attendMonth.filter(row => row.bonus && row.bonus[FIELD.type] === BONUS.month).length + 1,
+				count: attendMonth.filter(row => row.bonus?.[FIELD.type] === BONUS.month).length + 1,
 				desc: scheme.month.desc,
+			}
+			break;
+
+		/**
+		 * The Home scheme qualifies as a Bonus if the Member has already attended an '@Home' class earlier today
+		 */
+		case scheme.home
+			&& attendToday
+				.map(attend => attend.timetable[FIELD.id])				// get ID of the schedule attended
+				.map(id => source.client.schedule?.find(sched => sched[FIELD.id] === id) as ISchedule)
+				.filter(sched => sched && sched.location === BONUS.home)
+				.length >= scheme.home.level											// required number of Attends today
+			&& (isUndefined(elect) || elect !== BONUS.none):		// Member elected to not take Bonus
+
+			bonus = {
+				[FIELD.id]: scheme.home[FIELD.id],
+				[FIELD.type]: BONUS.home,
+				count: attendToday.length + 1,
+				desc: scheme.home.desc,
 			}
 			break;
 	}
 
-	if (upd.length)																		// whether or not we found an applicable Bonus
-		bonus.gift = upd;																//	let caller know about any pending Gift updates
+	if (upd.length)																					// whether or not we found an applicable Bonus
+		bonus.gift = upd;																			// let caller know about any pending Gift updates
 
 	return bonus;
 }
