@@ -3,25 +3,36 @@
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-  production: false,
+	production: true,
 
-  firebase: {
-    app: {
-      apiKey: "AIzaSyCQt465v9SVXfKE2gi5kHxWgExYyiXbQXQ",
-      authDomain: "whitefire-dev.firebaseapp.com",
-      databaseURL: "https://whitefire-dev.firebaseio.com",
-      projectId: "whitefire-dev",
-      storageBucket: "whitefire-dev.appspot.com",
-      messagingSenderId: "919087826501",
-    },
+	firebase: {
+		dev: {
+			apiKey: "AIzaSyC7TfX6HvGBNTOkDG3lMTjFt0u9WOIHNpQ",
+			authDomain: "whitefire-dev.firebaseapp.com",
+			databaseURL: "https://whitefire-dev.firebaseio.com",
+			projectId: "whitefire-dev",
+			storageBucket: "whitefire-dev.appspot.com",
+			messagingSenderId: "919087826501",
+			vapId: "BPYmYiHe66sCBPntd1hrG1-b7cEDSPqzGr_ULlSD55qhy1eS_Cw7Sm-8Sri7BXYPPDGkoXqhQprathL7k7HQVNQ",
+		},
+		prod: {
+			apiKey: "AIzaSyCYo4FvLtznrNtM-I73vzBjTSuyVXebqVI",
+			authDomain: "whitefire-51840.firebaseapp.com",
+			databaseURL: "https://whitefire-51840.firebaseio.com",
+			projectId: "whitefire-51840",
+			storageBucket: "whitefire-51840.appspot.com",
+			messagingSenderId: "241706183908",
+			appId: "1:241706183908:web:2e684f93d2af915e",
+			vapId: "BE6LYLfHnJUJXHXNuNPlNoEntZheAtqcyV-n2GYMCrdrtODf41vlVefVPmltvRfHuvwTE_4l9fLfy8W0baPikZA",
+		},
 
-    config: {
-      name: 'whiteFire',
-      region: 'us-central1',
-      automaticDataCollectionEnabled: true,
-    }
-  }
-};
+		config: {
+			name: 'whiteFire',
+			region: 'us-central1',
+			automaticDataCollectionEnabled: true,
+		}
+	}
+}
 
 /*
  * For easier debugging in development mode, you can import the following file
