@@ -6,7 +6,8 @@ import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 import { AppModule } from './app/app.module';
 import { environment } from './environments/environment';
 
-import 'hammerjs';
+// import 'hammerjs';
+// import * as Hammer from 'hammerjs';
 
 if (environment.production)
   enableProdMode();
